+++ NEWS.md
@@ -9,6 +9,8 @@
 * Save model objects via argument `file` in `add_ic` 
 after adding model fit criteria. (#478) 
 * Compute density ratios based on MCMC samples via `density_ratio`.
+* Ignore offsets in various post-processing methods via
+argument `offset`.
 
 ### Other changes